Update README.md
The establishment of police was forced by societal, demographic, and economic changes resulting from the urbanisation and expansion of the population during the industrial revolution. The British elite, in particular, were concerned about being victimised by the impoverished and unemployed in London. Sir Robert Peel, then-chief of the London Metropolitan Police, introduced the Metropolis Police Act in 1829 to restore "the social cohesion that was alleged to have been lost through urbanisation and industrialization." The police's principal goal is to keep the public safe from crime and disturbance.
The study is based on the use of open source data to map out crime, with a focus on burglary in Cleveland, Middlesbrough. Her Majesty's Inspectorate of Constabulary and Fire & Rescue Services (HMICERS) rated Cleveland Police as the first force in England and Wales to be rated inadequate across all areas of performance in the Police effectiveness, efficiency and legitimacy 2018/2019 (PEEL) report of 2019. In order to verify the veracity of this report, the study used the statistical tools of Orange, Pandas, and Tableau to address the spatial dimension of Burglars.
  The research through the study of environmental criminology observed the hotspot zone crime attractors are public facilities like shopping malls, high schools, taverns, convenience stores, apartment buildings and public housing projects. Different types of facilities increase or decrease crime in their immediate environment. Defensible space features of the built environment, demographics and, to some extent, the temporary environment are the major variables affecting the rates of crimes.
  Conclusively, the research recommended PANDA, SARA and SPATIAL crime models which are all models used to support problem-solving policing. The main focus of the models is the need to reduce crime in urban and suburban communities